backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==10.0.1
piexif==1.1.3

# Document Generation
reportlab==4.0.7
python-docx==0.8.11

# Async & File Handling
aiofiles==23.2.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Development (optional)
pytest==7.4.3
black==23.12.0
flake8==6.1.0